refactor(web): retire type back-imports from App.tsx (#642 phase 3, seam 5a pt 1)
KnobOpt, DesignAdvisory/DesignLoadError, and PatternData were exported from
App.tsx and import-type'd back into the components that actually own them
(ParamForm, AwaitingTrustPanel, ViewPanel/FarFieldChart). Move each type to
its natural home so nothing runtime-imports from App.tsx anymore — the
precondition for extracting DesignSession/TabStrip/contexts out of App.tsx
next.
